Return Ticker to libraries only for modularity.
squash! Return Ticker to libraries only for modularity.
