+++ CMakeLists.txt
@@ -32,7 +32,6 @@ set(CORE_SRCS
   cores/esp32/Schedule.cpp
   cores/esp32/Stream.cpp
   cores/esp32/StreamString.cpp
-  cores/esp32/Ticker.cpp
   cores/esp32/wiring_pulse.c
   cores/esp32/wiring_shift.c
   cores/esp32/WMath.cpp
